feat(claim-gates): extract fleet claim engines (GAP-462 Phase 1)
Move claim-drift into @revealui/claim-gates with parameterized root.
Thin scripts/validate/claim-drift.ts wrapper keeps capability-claims
local and product-runtime parity. Bump package/workspace/MIT counts
for the new workspace package.
