+++ apps/marketing/app/content/fair-source.ts
@@ -1,6 +1,6 @@
 // Sourced from: app/routes/FairSourcePage.tsx (Phase 1 extraction).
 // Hero headline + body math reference the METRICS license split
-// (validator: 23 MIT + 5 FSL + 2 internal = 30 packages). The internal
+// (validator: 24 MIT + 5 FSL + 2 internal = 31 packages). The internal
 // packages are @revealui/scripts and @revealui/apify-actor-governed-run
 // (both private, no license field).
 // Per the internal marketing-overhaul plan §4.4 + docs/MARKETING_METRICS.md §1.
