+++ docs/WHAT_WORKS_TODAY.md
@@ -130,7 +130,7 @@ cd revealui
 pnpm install
 pnpm gate                # Run the full CI gate locally
 pnpm test                # Run the full test suite
-pnpm typecheck:all       # Typecheck all 35 workspaces
+pnpm typecheck:all       # Typecheck all 36 workspaces
 pnpm validate:claims     # Run the marketing/docs claim-drift gate
 ```
