+++ packages/claim-gates/src/cli.ts
@@ -1,0 +1,56 @@
+import path from 'node:path';
+import { runClaimGates } from './run.js';
+import type { ClaimGatesCliOptions, ClaimGateResult, ClaimProfileName } from './types.js';
+
+function parseArgv(argv: readonly string[]): {
+  root?: string;
+  profile?: ClaimProfileName;
+  showFix: boolean;
+} {
+  let root: string | undefined;
+  let profile: ClaimProfileName | undefined;
+  let showFix = false;
+  for (let i = 0; i < argv.length; i++) {
+    const a = argv[i];
+    if (a === undefined) continue;
+    if (a === '--fix') showFix = true;
+    else if (a === '--root' && argv[i + 1]) {
+      root = argv[++i];
+    } else if (a.startsWith('--root=')) {
+      root = a.slice('--root='.length);
+    } else if (a === '--profile' && argv[i + 1]) {
+      profile = argv[++i] as ClaimProfileName;
+    } else if (a.startsWith('--profile=')) {
+      profile = a.slice('--profile='.length) as ClaimProfileName;
+    }
+  }
+  return { root, profile, showFix };
+}
+
+/**
+ * CLI entry used by the package bin and the revealui claim-drift wrapper.
+ * Does not handle --update-capability-baseline (revealui-local; wrapper only).
+ */
+export function runClaimGatesCli(options: ClaimGatesCliOptions): ClaimGateResult {
+  const argv = options.argv ?? process.argv;
+  const parsed = parseArgv(argv);
+  const root = path.resolve(options.root ?? parsed.root ?? process.cwd());
+  const result = runClaimGates({
+    root,
+    profile: options.profile ?? parsed.profile,
+    argv,
+    showFix: parsed.showFix,
+    capability: options.capability,
+  });
+  if (options.exit !== false) {
+    process.exit(result.exitCode);
+  }
+  return result;
+}
+
+const invoked = process.argv[1] ? path.resolve(process.argv[1]) : '';
+const selfJs = path.resolve(import.meta.dirname, 'cli.js');
+const selfTs = path.resolve(import.meta.dirname, 'cli.ts');
+if (invoked === selfJs || invoked === selfTs) {
+  runClaimGatesCli({ root: process.cwd(), argv: process.argv, exit: true });
+}
